Query alias internals

alias_expr(alias) returns the expression that was used to define an alias.

alias_env(alias) returns the environment in which the aliased expression is evaluated.

Usage

alias_expr(alias, envir = parent.frame())

alias_env(alias, envir = parent.frame())

Arguments

alias

the name of an alias, as a string

envir

the environment in which to look up the alias name (default: calling environment)

Value

alias_expr(alias) returns an unevaluated R expression (a name or a call).

alias_env(alias) returns an environment.